+++ Assets/CommonCoreModules/World/ObjectComponents/NavigationNode.cs
@@ -2,19 +2,19 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace CommonCore.Experimental
+namespace CommonCore.World
 {
 
     /// <summary>
     /// Path node for navigating NPCs or something
     /// </summary>
-    public class NavigationNodeEx : MonoBehaviour
+    public class NavigationNode : MonoBehaviour
     {
         public bool StartNode = false;
         public bool EndNode = false;
         public float DistanceThreshold = 2f;
-        public NavigationNodeEx PreviousNode;
-        public NavigationNodeEx NextNode;
+        public NavigationNode PreviousNode;
+        public NavigationNode NextNode;
 
     }
 }