+++ CHANGELOG.md
@@ -364,6 +364,8 @@
 * Added ActivatorFilter to ObjectActions
 * Added ContainsSpecific API to main playerflags source to check if it is in that source specifically
 * ViewModel now collides with ViewModel
+* Moved NavigationNodeEx out of experimental into World (as NavigationNode)
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]
+* Added NavigationNodeReachedTrigger action trigger [Untested]